Move display_document test to input_method test
